# API_Document/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "如意外卖接口文档",
    "version": "1.0",
    "title": "如意外卖"
  },
  "host": "localhost:9000",
  "basePath": "/",
...
  "paths": {
...
    "/order": {
      "put": {
        "tags": [
          "订单管理"
        ],
        "summary": "派送外卖订单",
        "operationId": "sendUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orders",
            "description": "orders",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Orders对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/again": {
      "post": {
        "tags": [
          "订单管理"
        ],
        "summary": "再来一单",
        "operationId": "againUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "order1",
            "description": "order1",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Orders对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/page": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "后台查看订单明细",
        "operationId": "pageUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "beginTime",
            "in": "query",
            "description": "beginTime",
            "required": false,
            "type": "string"
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "endTime",
            "required": false,
            "type": "string"
          },
          {
            "name": "number",
            "in": "query",
            "description": "number",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/submit": {
      "post": {
        "tags": [
          "订单管理"
        ],
        "summary": "用户下单",
        "operationId": "submitUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orders",
            "description": "orders",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Orders对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/order/userPage": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "订单分页数据查询",
        "operationId": "userPageUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
    "Orders对象": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "refType": null
        },
        "addressBookId": {
          "type": "integer",
          "format": "int64",
          "description": "地址id",
          "refType": null
        },
        "amount": {
          "type": "number",
          "description": "实收金额",
          "refType": null
        },
        "checkoutTime": {
          "type": "string",
          "format": "date-time",
          "description": "结账时间",
          "refType": null
        },
        "consignee": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "number": {
          "type": "string",
          "description": "订单号",
          "refType": null
        },
        "orderTime": {
          "type": "string",
          "format": "date-time",
          "description": "下单时间",
          "refType": null
        },
        "payMethod": {
          "type": "integer",
          "format": "int32",
          "description": "支付方式 1微信,2支付宝",
          "refType": null
        },
        "phone": {
          "type": "string",
          "refType": null
        },
        "remark": {
          "type": "string",
          "description": "备注",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "订单状态 1待付款，2待派送，3已派送，4已完成，5已取消",
          "refType": null
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "下单用户",
          "refType": null
        },
        "userName": {
          "type": "string",
          "refType": null
        }
      },
      "title": "Orders对象",
      "description": "订单表"
    },
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
...
  }
}